+++ inc/page_id_helper_functions.h
@@ -3,7 +3,7 @@
 
 #include<buffer_pool_man_types.h>
 
-unsigned long long int hash_page_id(PAGE_ID page_id);
+unsigned int hash_page_id(PAGE_ID page_id);
 
 int compare_page_id(PAGE_ID page_id1, PAGE_ID page_id2);
